refactor: extract stat summary helpers
